Rename persistent_config to persistent
