+++ src/lib.rs
@@ -1,3 +1,3 @@
 pub mod error;
-pub mod persistent_config;
+pub mod persistent;
 pub mod s3;
